# autoconfigure/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: S3 Viewer API
  version: 1.0.0
  description: Browse objects across multiple S3-compatible providers and buckets.
servers:
  - url: /
tags:
  - name: Viewer
paths:
  /s3-viewer/api/v1/providers:
    get:
      tags: [Viewer]
      operationId: listProviders
      summary: List configured S3 providers
      responses:
        "200":
          description: Providers returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderSummary"
  /s3-viewer/api/v1/providers/{providerId}/buckets:
    get:
      tags: [Viewer]
      operationId: listBuckets
      summary: List buckets for a provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: Buckets returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BucketSummary"
        "404":
          $ref: "#/components/responses/NotFound"
  /s3-viewer/api/v1/providers/{providerId}/buckets/{bucketName}/browse:
    get:
      tags: [Viewer]
      operationId: browseBucket
      summary: Browse objects inside a bucket path
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/BucketName"
        - name: path
          in: query
          required: false
          description: Prefix or virtual directory path to browse
          schema:
            type: string
      responses:
        "200":
          description: Directory listing returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrowseResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /s3-viewer/api/v1/providers/{providerId}/buckets/{bucketName}/download:
    get:
      tags: [Viewer]
      operationId: downloadObject
      summary: Download an object from a bucket
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/BucketName"
        - name: key
          in: query
          required: true
          description: Object key to download
          schema:
            type: string
      responses:
        "200":
          description: Object content streamed successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
  /s3-viewer/api/v1/providers/{providerId}/buckets/{bucketName}/upload:
    post:
      tags: [Viewer]
      operationId: uploadObject
      summary: Upload an object to a bucket path
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/BucketName"
        - name: path
          in: query
          required: false
          description: Target directory path (prefix) in the bucket
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
      responses:
        "201":
          description: Object uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectEntry"
        "404":
          $ref: "#/components/responses/NotFound"
  /s3-viewer/api/v1/providers/{providerId}/buckets/{bucketName}/objects:
    delete:
      tags: [Viewer]
      operationId: deleteObjects
      summary: Delete one or more objects from a bucket
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/BucketName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRequest"
      responses:
        "204":
          description: Objects deleted successfully
        "404":
          $ref: "#/components/responses/NotFound"
  /s3-viewer/api/v1/providers/{providerId}/buckets/{bucketName}/search:
    get:
      tags: [Viewer]
      operationId: searchObjects
      summary: Search objects in a bucket by name
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/BucketName"
        - name: query
          in: query
          required: true
          description: Search term matched against object names (case-insensitive)
          schema:
            type: string
        - name: path
          in: query
          required: false
          description: Restrict search to a specific prefix/directory
          schema:
            type: string
        - name: maxResults
          in: query
          required: false
          description: Maximum number of results to return (default 100)
          schema:
            type: integer
            format: int32
            default: 100
      responses:
        "200":
          description: Search results returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
    BucketName:
      name: bucketName
      in: path
      required: true
      schema:
        type: string
  responses:
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ProviderSummary:
      type: object
      required: [id, name, endpoint, region, bucketCount]
      properties:
        id:
          type: string
        name:
          type: string
        endpoint:
          type: string
        region:
          type: string
        bucketCount:
          type: integer
          format: int32
        pathStyleAccess:
          type: boolean
    BucketSummary:
      type: object
      required: [name, providerId]
      properties:
        name:
          type: string
        providerId:
          type: string
        objectCountHint:
          type: integer
          format: int64
        configured:
          type: boolean
    BrowseResponse:
      type: object
      required: [providerId, bucketName, path, entries]
      properties:
        providerId:
          type: string
        bucketName:
          type: string
        path:
          type: string
        parentPath:
          type: string
          nullable: true
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ObjectEntry"
    SearchResponse:
      type: object
      required: [providerId, bucketName, query, entries]
      properties:
        providerId:
          type: string
        bucketName:
          type: string
        query:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/ObjectEntry"
    ObjectEntry:
      type: object
      required: [name, key, type]
      properties:
        name:
          type: string
        key:
          type: string
        type:
          $ref: "#/components/schemas/ObjectEntryType"
        size:
          type: integer
          format: int64
          nullable: true
        lastModified:
          type: string
          format: date-time
          nullable: true
    ObjectEntryType:
      type: string
      enum: [DIRECTORY, FILE]
    DeleteRequest:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            type: string
          description: List of object keys to delete
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string